+++ src/printer/fdm/index.js
@@ -26,9 +26,7 @@ const updateHostname = (obj) => {
   const newHostname = () => {
     const hostnameLabel = document.getElementById("title-hostname");
     if (hostnameLabel) {
-      hostnameLabel.innerHTML = translate("glob.hostname") + ":";
-      document.getElementById("title-hostname").innerHTML =
-        context.version?.hostname || "localhost";
+      hostnameLabel.innerHTML = getHostname();
     }
   };
   const load = obj.load;
@@ -38,6 +36,11 @@ const updateHostname = (obj) => {
   };
   return obj;
 };
+
+const getHostname = () => context.version?.hostname || "localhost";
+const buildTitle = (title) => getHostname() + " - " +
+  (title.trim() || process.env.APP_NAME) +
+  " - " + process.env.APP_NAME;
 
 const updatePrinterStatus = (state) => {
   const query = { query: "#printer-status" };
@@ -61,11 +64,13 @@ const fdm = {
       path: "dashboard",
       html: require("../../views/dashboard.html"),
       module: updateHostname(dashboard),
+      getTitle: () => buildTitle(translate("home.link")),
     },
     {
       path: "temperature",
       html: require("../../views/temperature.html"),
       module: updateHostname(temperature),
+      getTitle: () => buildTitle(translate("temps.title")),
     },
     {
       path: "question",
@@ -84,6 +89,7 @@ const fdm = {
         path: "projects",
         html: require("../../views/projects.html"),
         module: updateHostname(projects),
+        getTitle: () => buildTitle(translate("proj.link")),
       }
       : null,
     process.env.WITH_SETTINGS ?
@@ -91,6 +97,7 @@ const fdm = {
         path: "settings",
         html: require("../../views/settings.html"),
         module: updateHostname(require("../components/settings.js").default),
+        getTitle: () => buildTitle(translate("settings.title")),
       }
       : null,
     process.env.WITH_CONTROLS ?
@@ -98,6 +105,7 @@ const fdm = {
         path: "control",
         html: require("../../views/control.html"),
         module: updateHostname(require("../components/control.js").default),
+        getTitle: () => buildTitle(translate("control.link")),
       } : null,
   ].filter(route => route != null),
   init: ({ version, printer, connection }) => {